refactor: move e2e tests to test/e2e
