+++ test/e2e/goHome.test.ts
@@ -1,7 +1,7 @@
 import { chromium, Page, Browser, BrowserContext, Cookie } from "playwright"
-import { hash } from "../src/node/util"
-import { CODE_SERVER_ADDRESS, PASSWORD, STORAGE } from "./constants"
-import { createCookieIfDoesntExist } from "./helpers"
+import { hash } from "../../src/node/util"
+import { CODE_SERVER_ADDRESS, PASSWORD, STORAGE, E2E_VIDEO_DIR } from "../utils/constants"
+import { createCookieIfDoesntExist } from "../utils/helpers"
 
 describe("go home", () => {
   let browser: Browser
@@ -45,7 +45,7 @@ describe("go home", () => {
 
     context = await browser.newContext({
       storageState: { cookies: maybeUpdatedCookies },
-      recordVideo: { dir: "./test/videos/" },
+      recordVideo: { dir: E2E_VIDEO_DIR },
     })
   })
